backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
cryptography==41.0.7
pycryptodome==3.19.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
PyJWT==2.8.0
gunicorn
